US 1040: fill out forms
This adds a routine to taxsolve_routines.c which accepts
the form structure and form data then outputs the xfdf format
file. This also includes the structures for the US1040,
schedule a, and schedule c. I've also added a script
which will download the forms and check the md5 sums on
the forms for sanity, then use pdftk to produce a new
pdf with the field populated.